Queue/debounce facade generation
